modules: Support any content in article panel
For the Blog template, we need more than just the regular
next article card, instead, we want to display 3 related
articles.

The Layout.ArticleStack module is limited to only 1 card
for this purpose, so add a new slot to Layout.ArticleStack
that allow us to put any type of content inside the panel.

https://phabricator.endlessm.com/T16699