contentGroup: Track existing context
HistoryStore item context is created within the ContentGroup
when clicking one of its cards. The context is defined as all
the models contained inside the group selection.

This behavior is desired most of the times, but now we do
have case where the group selection is irrelevant and should
keep track of the existing context instead.

Add a new property "track-context", to tell whether we want
to track the existing context or not.

https://phabricator.endlessm.com/T16699